test(map): bench the alignment kernel with divan instead of a hand-rolled loop
`examples/bench_align` timed the kernel by running a fixed iteration count and
dividing wall clock by the number of alignments. That reports one number per
run with no distribution behind it, and the number moved by up to 28% between
runs of the same binary on the same input -- wider than every change the
mapper has been getting, so the harness could not tell a real regression from
scheduler noise.

The same workload is now `benches/align.rs` under divan, which runs each case
to a time budget and reports fastest/median/mean. Three consecutive runs of the
anchored path, medians per case:

  read_len   run 1     run 2     run 3     spread
  51  0mm    5.624 us  5.499 us  5.666 us   3.0%
  100 0mm    9.082 us  9.895 us  9.624 us   8.9%
  150 0mm    11.52 us  11.33 us  11.54 us   1.9%
  51  2mm    5.957 us  5.812 us  5.958 us   2.5%
  100 2mm    10.10 us  10.24 us  10.54 us   4.4%
  150 2mm    11.95 us  11.79 us  12.12 us   2.8%

Against the old loop's 21.7 / 24.8 / 27.8 ns per alignment for the 51bp
zero-mismatch case across three runs.

Cases are now selectable (`cargo bench -p salmon-map -- anchored`, or by read
length) instead of the binary always running all twelve, and the item counter
reports alignments per second directly. The generator and the two `AlignConfig`
paths are unchanged, so the workload is the same one the example measured.
